Convert to local datetime during timezone conversion
When we convert a DateTime to another timezone, there could be a case
that the daylight savings starts or ends during the course of that
day. In this case, the changed daylight savings time should apply only
after that point in the day. This is achieved by converting the
date-time object to a local-date-time object and then converting it
back to a date-time-object.

For example, in Asia/Amman, daylight savings ends on 26 Oct at 1 am,
and the timezone offset becomes +03:00 from +02:00.
;; Older behaviour:
(let [dtz ^DateTimeZone (clj-time.core/time-zone-for-id "Asia/Amman")]
  (-> (clj-time.core/date-time 2018 3 26)
      (clj-time.core/from-time-zone dtz) str))
;; => 2018-10-26T00:00:00.000+02:00

The above should not be in the +02:00 timezone, but in the +03:00
timezone
;; Altered behaviour:
(let [dtz ^DateTimeZone (clj-time.core/time-zone-for-id "Asia/Amman")]
  (-> (clj-time.core/date-time 2018 10 26)
      (clj-time.core/from-time-zone dtz)))
;; => 2018-10-26T00:00:00.000+03:00